requirements.txt
# faiss-cpu 1.9.x has no wheels for Python 3.12+; 1.13.x supports current macOS/arm64 builds.
streamlit==1.41.0
pandas==2.2.3
openpyxl==3.1.5
plotly==5.24.1
numpy>=2.1.3,<3
scikit-learn==1.6.0
faiss-cpu==1.13.2
openai==1.57.0
langchain==0.3.12
langchain-openai==0.2.12
langchain-community==0.3.12
pdfplumber==0.11.4
tiktoken>=0.12.0
python-dotenv==1.0.1
